+++ providers/index.js
@@ -1,6 +1,0 @@
-module.exports = {
-  esri: require('./esri'),
-  bing: require('./bing'),
-  openstreetmap: require('./openstreetmap'),
-  toporama: require('./toporama')
-}